@@ -1,5 +1,4 @@
-/** @type {import('tailwindcss').Config} */
-const colors = require('tailwindcss/colors')
+import type { Config } from 'tailwindcss';
 
 export default {
   content: ['./src/**/*.{html,js,svelte,ts}'],
@@ -16,5 +15,5 @@ export default {
     },
   },
   plugins: [],
-}
+} satisfies Config;
 
